pyproject.toml
[project]
name = "novel-writer-master"
version = "0.1.0"
description = "AI-assisted novel writing tool with built-in anti-AI-detection engine"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.40.0",
    "openai>=1.60.0",
    "jieba>=0.42.1",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
vector = ["chromadb>=0.5.0"]
export = ["python-docx>=1.1.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0", "ruff>=0.8.0"]

[project.scripts]
novel-writer-master = "novel_writer.ui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/novel_writer"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
